Test and statistics guide update and test deletion update
Broken test_stat_guide.php into test_guide.php and stat_guide.php and overhauled them

Updated guide.php, admin_guide.php, professor_guide.php and custom.css to reflect those changes

Added a message that informs the user that it is not possible to delete a test with results

Small fixes in test_stats.php

+++ src/test/test_delete.php
@@ -4,14 +4,33 @@ include $_SERVER['DOCUMENT_ROOT']."/libraries/general.php";
 chk_access(0);
 connect();
 
-$del_st = prepare_stmt("DELETE FROM TEST WHERE id_test=?");
-$del_st->bind_param("i", $_GET['id']);
-execute_stmt($del_st);
-$del_st->close();
+$chk_st = prepare_stmt("SELECT COUNT(*) AS n FROM PROVE WHERE fk_test=?");
+$chk_st->bind_param("i", $_GET['id']);
+$ret = execute_stmt($chk_st);
+$chk_st->close();
 
-writelog("Test ".$_GET['id']." cancellato");
+$r = $ret->fetch_assoc();
 
-$_SESSION['alert'] = "Test eliminato correttamente";
-header("Location: /test/test.php");
+// If some results are present the deletion is blocked
+if($r['n'] === 0)
+{
+    $del_st = prepare_stmt("DELETE FROM TEST WHERE id_test=?");
+    $del_st->bind_param("i", $_GET['id']);
+    execute_stmt($del_st);
+    $del_st->close();
+
+    writelog("Test ".$_GET['id']." cancellato");
+
+    $_SESSION['alert'] = "Test eliminato correttamente";
+    header("Location: /test/test.php");
+}
+else
+{
+    writelog("Tentativo cancellazione test ".$_GET['id']." bloccato; esistono ".$r['n']." prove");
+
+    $_SESSION['alert'] = "Impossibile eliminare il test: esistono ".$r['n']." prove associate";
+    header("Location: /test/test_modify.php?id=".$_GET['id']);
+}
+
 exit;
 ?>
